response is fixed

+++ src/main/java/com/yash/ytms/controller/JobController.java
@@ -9,6 +9,7 @@ package com.yash.ytms.controller;
  * @since 29-05-2024
  */
 import com.yash.ytms.domain.Job;
+import com.yash.ytms.dto.ResponseWrapperDto;
 import com.yash.ytms.exception.ApplicationException;
 import com.yash.ytms.services.IServices.JobService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +18,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/jobs")
@@ -26,15 +28,20 @@ public class JobController {
     private JobService jobService;
 
     @GetMapping
-    public List<Job> getAllJobs() {
-        return jobService.findAll();
+    public List<Job>  getAllJobs() {
+        List<Job>  jobs = jobService.findAll();
+         return  jobs;
+
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Job> getJobById(@PathVariable Long id) {
-        return jobService.findById(id)
-                .map(job -> ResponseEntity.ok().body(job))
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<ResponseWrapperDto> getJobById(@PathVariable Long id) {
+        Job job = jobService.findById(id);
+        ResponseWrapperDto wrapperDto = new ResponseWrapperDto();
+        wrapperDto.setData(job);
+        wrapperDto.setStatus("SUCCESS");
+        return  new ResponseEntity<>(wrapperDto,HttpStatus.OK);
+
     }
 
     @PostMapping("/save")
@@ -45,21 +52,20 @@ public class JobController {
 
     @PutMapping("/jobs/{id}")
     public ResponseEntity<Job> updateJob(@PathVariable Long id, @RequestBody Job jobDetails) {
-        return jobService.findById(id)
-                .map(job -> {
+        Optional<Job> jobOptional = Optional.ofNullable(jobService.findById(id));
+        Job job= jobOptional.get();
                     job.setJobRole(jobDetails.getJobRole());
                     job.setJobLocation(jobDetails.getJobLocation());
                     job.setCompany(jobDetails.getCompany());
                     job.setStatus(jobDetails.getStatus());
-                    return ResponseEntity.ok().body(jobService.save(job));
-                })
-                .orElse(ResponseEntity.notFound().build());
+                     jobService.save(job);
+                     return new ResponseEntity<>(HttpStatus.OK);
+
     }
 
     @PatchMapping("/{id}")
-    public ResponseEntity<Job> updateJobStatus(@PathVariable Long id, @RequestBody Job jobDetails) {
-        Job job = jobService.findById(id)
-                .orElseThrow(() -> new ApplicationException("Job not found with id " + id));
+    public ResponseEntity<ResponseWrapperDto> updateJobStatus(@PathVariable Long id, @RequestBody Job jobDetails) {
+        Job job = Optional.ofNullable(jobService.findById(id)).get();
         job.setStatus(jobDetails.getStatus());
         job.setJobLocation(jobDetails.getJobLocation());
         job.setJobRole(jobDetails.getJobRole());
@@ -68,22 +74,30 @@ public class JobController {
         job.setJobListingUrl(jobDetails.getJobListingUrl());
         job.setJobDescription(jobDetails.getJobDescription());
         job.setSalary(jobDetails.getSalary());
-        Job updatedJob = jobService.save(job);
-        return ResponseEntity.ok(updatedJob);
+        ResponseWrapperDto updatedJob = jobService.save(job);
+        return new ResponseEntity(updatedJob, HttpStatus.OK);
     }
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deleteJob(@PathVariable Long id) {
-        return jobService.findById(id)
-                .map(job -> {
-                    jobService.deleteById(id);
-                    return ResponseEntity.ok().build();
-                })
-                .orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<ResponseWrapperDto> deleteJob(@PathVariable Long id) {
+        ResponseWrapperDto wrapperDto= new ResponseWrapperDto();
+        Job job = jobService.findById(id);
+        if(job!=null){
+            jobService.deleteById(id);
+        }
+                wrapperDto.setStatus("SUCCESS");
+        wrapperDto.setMessage("Deleted job successfully");
+        return new ResponseEntity<>(wrapperDto,HttpStatus.OK);
     }
 
     @GetMapping("/status/{status}")
-    public List<Job> getJobsByStatus(@PathVariable String status) {
-        return jobService.findByStatus(status);
+    public ResponseEntity getJobsByStatus(@PathVariable String status) {
+        List<Job> jobs = jobService.findByStatus(status);
+        ResponseWrapperDto wrapperDto = new ResponseWrapperDto();
+        wrapperDto.setData(jobs);
+        wrapperDto.setStatus("SUCCESS");
+        return  new ResponseEntity<>(wrapperDto,HttpStatus.OK);
+
+
     }
 }
